@@ -1,28 +1,35 @@
 {
-  "name": "aditya-magnets",
+  "name": "aaditya_magnets_rr",
   "private": true,
-  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "build": "react-router build",
+    "dev": "react-router dev",
+    "start": "react-router-serve ./build/server/index.js",
+    "typecheck": "react-router typegen && tsc"
   },
   "dependencies": {
-    "next": "14.2.5",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "next-seo": "6.5.0",
-    "clsx": "2.1.1"
+    "@react-router/node": "^7.7.1",
+    "@react-router/serve": "^7.7.1",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "isbot": "^5.1.27",
+    "lucide-react": "^0.542.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^7.7.1",
+    "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
-    "@types/node": "20.11.30",
-    "@types/react": "18.2.74",
-    "@types/react-dom": "18.2.24",
-    "typescript": "5.4.5",
-    "eslint": "8.57.0",
-    "eslint-config-next": "14.2.5",
-    "tailwindcss": "3.4.7",
-    "autoprefixer": "10.4.14"
+    "@react-router/dev": "^7.7.1",
+    "@tailwindcss/vite": "^4.1.4",
+    "@types/node": "^20",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "tailwindcss": "^4.1.4",
+    "tw-animate-css": "^1.3.7",
+    "typescript": "^5.8.3",
+    "vite": "^6.3.3",
+    "vite-tsconfig-paths": "^5.1.4"
   }
 }
